requirements.txt
streamlit
torch
torchvision
transformers
Pillow
numpy>=2.0.0
opencv-python
matplotlib
scikit-learn